comment on networksize and less size in protocolgame npctrade

+++ src/protocolgame.cpp
@@ -1029,7 +1029,7 @@ void ProtocolGame::parsePlayerPurchase(NetworkMessage& msg)
 	uint8_t amount = msg.getByte();
 	bool ignoreCap = msg.getByte() != 0;
 	bool inBackpacks = msg.getByte() != 0;
-	uint16_t specialId = msg.get<uint16_t>();
+	uint8_t specialId = msg.get<uint8_t>();
 
 	addGameTaskTimed(DISPATCHER_TASK_EXPIRATION, &Game::playerPurchaseItem, player->getID(), id, count, amount, ignoreCap, inBackpacks, specialId);
 }
@@ -1040,7 +1040,7 @@ void ProtocolGame::parsePlayerSale(NetworkMessage& msg)
 	uint8_t count = msg.getByte();
 	uint8_t amount = msg.getByte();
 	bool ignoreEquipped = msg.getByte() != 0;
-	uint16_t specialId = msg.get<uint16_t>();
+	uint8_t specialId = msg.get<uint8_t>();
 	addGameTaskTimed(DISPATCHER_TASK_EXPIRATION, &Game::playerSellItem, player->getID(), id, count, amount, ignoreEquipped, specialId);
 }
